moved cmm stuff into ccm class
