+++ src/main/resources/org/jpeek/metrics/CCM.xsl
@@ -62,6 +62,9 @@ SOFTWARE.
     </xsl:variable>
     <xsl:copy>
       <xsl:variable name="nc" select="count($edges/edge)"/>
+      <!--
+      @todo add capture ncc value from skeleton
+      -->
       <xsl:variable name="ncc" select="count(distinct-values($edges/edge/method/text()))"/>
       <xsl:variable name="nmp" select="(count($methods) * (count($methods) - 1)) div 2"/>
       <xsl:attribute name="value">
